pyproject.toml
[tool.poetry]
name = "ingest-nyc-taxi-data"
version = "0.1.0"
description = ""
authors = ["slchangtw <shunlung.chang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.1"
pyarrow = "^13.0.0"
sqlalchemy = "<2.0"
pydantic-settings = {extras = ["dotenv"], version = "^2.0.3"}
prefect = "^2.14.2"
python-dateutil = "^2.8.2"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
notebook = "^7.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
